requirements.txt
asgiref==3.3.4
Django==3.2.3
django-modern-rpc==0.12.1
pytz==2021.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
